Prevent dark mode flicker
